[#201] refactor(chat): ChatRoomService 제거
- 관련 수정 파일들까지 모두 수정하여 테스트 확인 완료
