+++ src/test/java/com/dduru/gildongmu/chat/service/GroupChatRoomServiceTest.java
@@ -57,7 +57,7 @@ class GroupChatRoomServiceTest {
 
     @Test
     @DisplayName("초대 대상이 모두 존재하지 않으면 응답만 반환하고 멤버 조회 쿼리를 실행하지 않는다")
-    void inviteMembersToGroupRoom_allInviteesMissing_returnsEmptyWithoutMemberLookup() {
+    void inviteMembers_allInviteesMissing_returnsEmptyWithoutMemberLookup() {
         Long roomId = 10L;
         Long hostId = 1L;
         GroupChatInviteRequest request = new GroupChatInviteRequest(List.of(101L, 102L));
@@ -67,7 +67,7 @@ class GroupChatRoomServiceTest {
         when(userRepository.findAllById(request.inviteeUserIds()))
                 .thenReturn(List.of());
 
-        GroupChatInviteResponse response = groupChatRoomService.inviteMembersToGroupRoom(hostId, roomId, request);
+        GroupChatInviteResponse response = groupChatRoomService.inviteMembers(hostId, roomId, request);
 
         assertThat(response.roomId()).isEqualTo(roomId);
         assertThat(response.addedMemberCount()).isZero();
@@ -80,7 +80,7 @@ class GroupChatRoomServiceTest {
 
     @Test
     @DisplayName("존재하는 사용자와 없는 사용자가 섞여 있으면 유효한 사용자만 초대하고 누락 목록을 응답에 담는다")
-    void inviteMembersToGroupRoom_mixedInvitees_returnsPartialSuccess() {
+    void inviteMembers_mixedInvitees_returnsPartialSuccess() {
         Long roomId = 10L;
         Long hostId = 1L;
         Long validUserId = 2L;
@@ -97,7 +97,7 @@ class GroupChatRoomServiceTest {
                 .thenReturn(List.of());
         when(chatRoomMemberRepository.countByRoom(chatRoom)).thenReturn(1);
 
-        GroupChatInviteResponse response = groupChatRoomService.inviteMembersToGroupRoom(hostId, roomId, request);
+        GroupChatInviteResponse response = groupChatRoomService.inviteMembers(hostId, roomId, request);
 
         assertThat(response.roomId()).isEqualTo(roomId);
         assertThat(response.addedMemberCount()).isEqualTo(1);
@@ -129,7 +129,7 @@ class GroupChatRoomServiceTest {
         when(chatRoomMemberRepository.findExistingUserIdsByRoomIdAndUserIdIn(roomId, Set.of(existingMemberId)))
                 .thenReturn(List.of(existingMemberId));
 
-        GroupChatInviteResponse response = groupChatRoomService.inviteMembersToGroupRoom(hostId, roomId, request);
+        GroupChatInviteResponse response = groupChatRoomService.inviteMembers(hostId, roomId, request);
 
         assertThat(response.addedMemberCount()).isZero();
         assertThat(response.missingUserIds()).isEmpty();
@@ -140,27 +140,27 @@ class GroupChatRoomServiceTest {
 
     @Test
     @DisplayName("요청자가 게시글 작성자가 아니면 초대를 거부한다")
-    void inviteMembersToGroupRoom_nonHost_throwsAccessDenied() {
+    void inviteMembers_nonHost_throwsAccessDenied() {
         Long roomId = 10L;
         GroupChatInviteRequest request = new GroupChatInviteRequest(List.of(2L));
 
         when(chatRoomRepository.findByIdAndRoomTypeWithPostUser(roomId, ChatRoomType.GROUP))
                 .thenReturn(Optional.of(activeGroupRoom(roomId, 1L, 4)));
 
-        assertThatThrownBy(() -> groupChatRoomService.inviteMembersToGroupRoom(99L, roomId, request))
+        assertThatThrownBy(() -> groupChatRoomService.inviteMembers(99L, roomId, request))
                 .isInstanceOf(GroupChatRoomInviteAccessDeniedException.class);
     }
 
     @Test
     @DisplayName("종료된 그룹 채팅방에는 초대할 수 없다")
-    void inviteMembersToGroupRoom_closedRoom_throwsBusinessException() {
+    void inviteMembers_throwsBusinessException() {
         Long roomId = 10L;
         GroupChatInviteRequest request = new GroupChatInviteRequest(List.of(2L));
 
         when(chatRoomRepository.findByIdAndRoomTypeWithPostUser(roomId, ChatRoomType.GROUP))
                 .thenReturn(Optional.of(groupRoom(roomId, 1L, 4, ChatRoomStatus.CLOSED)));
 
-        assertThatThrownBy(() -> groupChatRoomService.inviteMembersToGroupRoom(1L, roomId, request))
+        assertThatThrownBy(() -> groupChatRoomService.inviteMembers(1L, roomId, request))
                 .isInstanceOf(BusinessException.class)
                 .extracting("errorCode")
                 .isEqualTo(ErrorCode.CHAT_ROOM_CLOSED);
@@ -168,7 +168,7 @@ class GroupChatRoomServiceTest {
 
     @Test
     @DisplayName("정원을 초과하면 예외가 발생한다")
-    void inviteMembersToGroupRoom_capacityExceeded_throwsException() {
+    void inviteMembers_capacityExceeded_throwsException() {
         Long roomId = 10L;
         Long hostId = 1L;
         Long inviteeId = 2L;
@@ -184,7 +184,7 @@ class GroupChatRoomServiceTest {
                 .thenReturn(List.of());
         when(chatRoomMemberRepository.countByRoom(chatRoom)).thenReturn(1);
 
-        assertThatThrownBy(() -> groupChatRoomService.inviteMembersToGroupRoom(hostId, roomId, request))
+        assertThatThrownBy(() -> groupChatRoomService.inviteMembers(hostId, roomId, request))
                 .isInstanceOf(ChatRoomCapacityExceededException.class);
     }
 
@@ -204,6 +204,88 @@ class GroupChatRoomServiceTest {
                 .build();
         ReflectionTestUtils.setField(chatRoom, "id", roomId);
         return chatRoom;
+    }
+
+/*
+    @Test
+    @DisplayName("그룹 채팅 초대 요청은 GroupChatRoomService에 위임한다")
+    void inviteMembersToGroupRoom_delegatesToGroupChatRoomService() {
+        GroupChatInviteRequest request = new GroupChatInviteRequest(java.util.List.of(2L, 3L));
+        GroupChatInviteResponse expected = new GroupChatInviteResponse(5L, 1, java.util.List.of(3L), java.util.List.of());
+
+        when(groupChatRoomService.inviteMembersToGroupRoom(1L, 5L, request)).thenReturn(expected);
+
+        GroupChatInviteResponse actual = groupChatRoomService.inviteMembersToGroupRoom(1L, 5L, request);
+
+        assertThat(actual).isEqualTo(expected);
+        verify(groupChatRoomService).inviteMembersToGroupRoom(1L, 5L, request);
+    }
+*/
+
+    @Test
+    @DisplayName("게시글용 그룹 채팅방 생성 시 방을 저장하고 작성자를 HOST로 추가한다")
+    void createPendingRoomForPost_savesRoomAndHostMember() {
+        Post post = Post.builder()
+                .title("제목")
+                .recruitCapacity(3)
+                .build();
+        ReflectionTestUtils.setField(post, "id", 10L);
+        User author = user(1L);
+
+        when(chatRoomRepository.save(org.mockito.ArgumentMatchers.any(ChatRoom.class))).thenAnswer(invocation -> {
+            ChatRoom savedRoom = invocation.getArgument(0);
+            ReflectionTestUtils.setField(savedRoom, "id", 20L);
+            return savedRoom;
+        });
+
+        groupChatRoomService.createPendingRoomForPost(post, author);
+
+        ArgumentCaptor<ChatRoom> roomCaptor = ArgumentCaptor.forClass(ChatRoom.class);
+        verify(chatRoomRepository).save(roomCaptor.capture());
+        ChatRoom savedRoom = roomCaptor.getValue();
+        assertThat(savedRoom.getRoomType()).isEqualTo(ChatRoomType.GROUP);
+        assertThat(savedRoom.getStatus()).isEqualTo(ChatRoomStatus.PENDING);
+        assertThat(savedRoom.getMaxCapacity()).isEqualTo(3);
+
+        ArgumentCaptor<ChatRoomMember> memberCaptor = ArgumentCaptor.forClass(ChatRoomMember.class);
+        verify(chatRoomMemberRepository).save(memberCaptor.capture());
+        ChatRoomMember savedMember = memberCaptor.getValue();
+        assertThat(savedMember.getRoom()).isSameAs(savedRoom);
+        assertThat(savedMember.getUser()).isSameAs(author);
+        assertThat(savedMember.getRole()).isEqualTo(ChatMemberRole.HOST);
+    }
+
+    @Test
+    @DisplayName("첫 메시지 이후 그룹 채팅방은 PENDING에서 ACTIVE로 전환된다")
+    void activateGroupChatOnFirstMessage_activatesPendingRoom() {
+        ChatRoom groupRoom = ChatRoom.builder()
+                .roomType(ChatRoomType.GROUP)
+                .status(ChatRoomStatus.PENDING)
+                .maxCapacity(4)
+                .build();
+
+        when(chatRoomRepository.getByIdOrThrow(1L)).thenReturn(groupRoom);
+
+         groupChatRoomService.activateChatOnFirstMessage(1L);
+
+        assertThat(groupRoom.getStatus()).isEqualTo(ChatRoomStatus.ACTIVE);
+    }
+
+    @Test
+    @DisplayName("그룹 채팅방이 아니면 첫 메시지 처리 시 상태를 변경하지 않는다")
+    void activateGroupChatOnFirstMessage_nonRoomKeepsStatus() {
+        ChatRoom privateRoom = ChatRoom.builder()
+                .roomType(ChatRoomType.PRIVATE)
+                .status(ChatRoomStatus.ACTIVE)
+                .maxCapacity(2)
+                .build();
+
+        when(chatRoomRepository.getByIdOrThrow(1L)).thenReturn(privateRoom);
+
+        groupChatRoomService.activateChatOnFirstMessage(1L);
+
+        assertThat(privateRoom.getStatus()).isEqualTo(ChatRoomStatus.ACTIVE);
+        verify(chatRoomMemberRepository, never()).save(org.mockito.ArgumentMatchers.any(ChatRoomMember.class));
     }
 
     private static User user(Long userId) {
